+++ ACBrFramework/ACBrFramework.Net/ACBrFramework.Net/ACBrPAF/Registros/ACBrPAFRegistroR3.cs
@@ -4,7 +4,7 @@ namespace ACBrFramework
 	{
 		public string TOT_PARCIAL { get; set; }
 
-		public double VL_ACUM { get; set; }
+		public decimal VL_ACUM { get; set; }
 
 		public bool RegistroValido { get; set; }
 	}
